# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meken API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/config": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Config",
        "operationId": "config_v1_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfiguration"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Delay in seconds.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/search": {
      "post": {
        "tags": [
          "discovery"
        ],
        "summary": "Search",
        "operationId": "search_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingPage"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Delay in seconds.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/apartments/{apartment_id}": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Detail",
        "operationId": "detail_v1_apartments__apartment_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "apartment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Apartment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Listing"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Delay in seconds."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/apartments/{apartment_id}/verify": {
      "post": {
        "tags": [
          "discovery"
        ],
        "summary": "Verify",
        "operationId": "verify_v1_apartments__apartment_id__verify_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "apartment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Apartment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Delay in seconds."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/favorites": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Favorites",
        "operationId": "favorites_v1_favorites_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingPage"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Delay in seconds.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/conversations/{conversation_id}/preferences": {
      "put": {
        "tags": [
          "discovery"
        ],
        "summary": "Update Preferences",
        "operationId": "update_preferences_v1_conversations__conversation_id__preferences_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preferences"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Delay in seconds."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "APIErrorResponse"
      },
...
      "AppConfiguration": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ],
            "title": "Mode"
          },
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled"
          },
          "cities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cities"
          },
          "privacy_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Url"
          },
          "terms_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Url"
          },
          "retention_days": {
            "type": "integer",
            "title": "Retention Days"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities",
            "description": "Supported optional API features; absent means legacy."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "ai_enabled",
          "cities",
          "privacy_url",
          "terms_url",
          "retention_days"
        ],
        "title": "AppConfiguration"
      },
...
      "SearchRequest": {
        "properties": {
          "preferences": {
            "$ref": "#/components/schemas/Preferences"
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 20
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SearchRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}